fixed invalid decomposable reward item ids and implemented a startup check to avoid this in the future

+++ game-server/src/com/aionemu/gameserver/controllers/attack/KillList.java
@@ -35,7 +35,7 @@ public class KillList {
 
 		synchronized (killTimes) {
 			for (Iterator<Long> i = killTimes.iterator(); i.hasNext();) {
-				if (now - i.next().longValue() > CustomConfig.PVP_DAY_DURATION) {
+				if (now - i.next() > CustomConfig.PVP_DAY_DURATION) {
 					i.remove();
 				} else {
 					killCount++;
